Unification Step 4: merge frame contexts into one PresetContext (full f64 time/beat)
EffectContext + GeneratorContext collapse into one PresetContext (union of
both fields). Resolves the f32/f64 fork by carrying time/beat as f64 from the
playback clock all the way to the single GPU uniform-write boundary -- no
f32 round-trip -- so beat phase stays exact over long shows. CompositorFrame
time/beat widened to f64; redundant intermediate conversions removed.
generator_context.rs deleted; no type-alias shim. Parity value-neutral
(byte-identical output vs prior HEAD).

+++ crates/manifold-renderer/src/compositor.rs
@@ -25,8 +25,8 @@ pub struct CompositeLayerDescriptor<'a> {
 
 /// Frame context passed to the compositor each tick.
 pub struct CompositorFrame<'a> {
-    pub time: f32,
-    pub beat: f32,
+    pub time: f64,
+    pub beat: f64,
     pub dt: f32,
     pub frame_count: u64,
     pub compositor_dirty: bool,
